Module Odoc_parser.Warning

Warnings produced during parsing.

type t = {

location : Loc.span;

message : string;

}

Warnings are represented as record containing the human-readable text of the warning alongside the location of the offending text in the source

val pp : Format.formatter -> t -> unit

Pretty printer for t

val to_string : t -> string

to_string will format the location and warning as text to be printed.